Split sql.go into smaller files per type
